if set_time is provided, use it instead of parsing
If set_time is provided, then it will be used and the code will not
attempt to parse the description line to find time in it. This
assumes that the user is setting time deliberately, so it should
override any other logic.

In case set_time is not provided, then parsing will be attempted but if
it fails, instead of just printing a warning, the code should throw an
error with the error message so that the code stops continuing further
and alerts the user about the issue.