Parsing time in EFIT header (#8)
* Try harder to parse time out of EFIT header and giveup with a warning message if that still fails.

* Add set_time option to readg() as backup for failing to parse time in geqdsk

* if set_time is provided, use it instead of parsing

If set_time is provided, then it will be used and the code will not
attempt to parse the description line to find time in it. This
assumes that the user is setting time deliberately, so it should
override any other logic.

In case set_time is not provided, then parsing will be attempted but if
it fails, instead of just printing a warning, the code should throw an
error with the error message so that the code stops continuing further
and alerts the user about the issue.

* Broken (needs fix and cleanup): split header parsing into its own file and add test for header parsing

* parse header testing works now

The separated parse header function now works and is tested. To add
more tests, the header list can be expanded with other cases and
expected outcomes.

---------

Co-authored-by: Anchal Gupta <guptaa@fusion.gat.com>